Fix up the indexer so that we don't commit after every change.
This sets both a timeout and a mutation threshold for the index commit
(5s and 1000 respecitvely) to make sure we're not commiting after
every single change, which would slow things right down on a large
number of change events.

There may be a better solution to this with recv_deadline, however
that is still in nightly.